Stats refactor for precision & clarity

+++ src/main/java/com/example/electricity_backend/dto/Price.java
@@ -5,7 +5,7 @@ import org.springframework.lang.NonNull;
 
 public record Price(
     @NonNull String timestamp,
-    float value,
+    double value,
     @NonNull String resolutionMinutes
 ) {}
